% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENIC function.R
\name{ExportBasicLoom}
\alias{ExportBasicLoom}
\title{FUNCTION_TITLE}
\usage{
ExportBasicLoom(seu, dir = getwd(), file.name = "seu.loom")
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{dir}{PARAM_DESCRIPTION, Default: getwd()}

\item{file.name}{PARAM_DESCRIPTION, Default: 'seu.loom'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
