% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanger.R
\name{RunCellranger}
\alias{RunCellranger}
\title{FUNCTION_TITLE}
\usage{
RunCellranger(
  fastqs,
  IDs,
  spe,
  program.path = NULL,
  ref = NULL,
  walltime = NULL,
  project = NULL,
  nodes = NULL,
  ppn = NULL,
  email = NULL,
  sequential = F
)
}
\arguments{
\item{fastqs}{PARAM_DESCRIPTION}

\item{IDs}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION}

\item{program.path}{PARAM_DESCRIPTION, Default: NULL}

\item{ref}{PARAM_DESCRIPTION, Default: NULL}

\item{walltime}{PARAM_DESCRIPTION, Default: NULL}

\item{project}{PARAM_DESCRIPTION, Default: NULL}

\item{nodes}{PARAM_DESCRIPTION, Default: NULL}

\item{ppn}{PARAM_DESCRIPTION, Default: NULL}

\item{email}{PARAM_DESCRIPTION, Default: NULL}

\item{sequential}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
