% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunRscript.R
\name{RunRscript}
\alias{RunRscript}
\title{FUNCTION_TITLE}
\usage{
RunRscript(file, nCore = 36)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}

\item{nCore}{PARAM_DESCRIPTION, Default: 36}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
