% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENIC function.R
\name{RunScenic}
\alias{RunScenic}
\title{FUNCTION_TITLE}
\usage{
RunScenic(
  loom.path,
  spe = "human",
  project.name = "Scenic_project",
  dir = getwd(),
  nCore = 36
)
}
\arguments{
\item{loom.path}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION, Default: 'human'}

\item{project.name}{PARAM_DESCRIPTION, Default: 'Scenic_project'}

\item{dir}{PARAM_DESCRIPTION, Default: getwd()}

\item{nCore}{PARAM_DESCRIPTION, Default: 36}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
