% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vscjob_CreateConfig}
\alias{vscjob_CreateConfig}
\title{FUNCTION_TITLE}
\usage{
vscjob_CreateConfig(force = FALSE)
}
\arguments{
\item{force}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
