% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vscjob_SetConfig}
\alias{vscjob_SetConfig}
\title{FUNCTION_TITLE}
\usage{
vscjob_SetConfig(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
