% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check.sys.dir.exist}
\alias{check.sys.dir.exist}
\title{FUNCTION_TITLE}
\usage{
check.sys.dir.exist(x, nm = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{nm}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
