% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunRscript.R
\name{RunRscript}
\alias{RunRscript}
\title{FUNCTION_TITLE}
\usage{
RunRscript(
  file,
  walltime = NULL,
  project = NULL,
  nodes = NULL,
  ppn = NULL,
  email = NULL
)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}

\item{walltime}{PARAM_DESCRIPTION, Default: NULL}

\item{project}{PARAM_DESCRIPTION, Default: NULL}

\item{nodes}{PARAM_DESCRIPTION, Default: NULL}

\item{ppn}{PARAM_DESCRIPTION, Default: NULL}

\item{email}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
