% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENIC function.R
\name{RunScenic}
\alias{RunScenic}
\title{FUNCTION_TITLE}
\usage{
RunScenic(
  loom.path,
  spe,
  project.name = "Scenic_project",
  dir = getwd(),
  walltime = NULL,
  project = NULL,
  nodes = NULL,
  ppn = NULL,
  email = NULL
)
}
\arguments{
\item{loom.path}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION}

\item{project.name}{PARAM_DESCRIPTION, Default: 'Scenic_project'}

\item{dir}{PARAM_DESCRIPTION, Default: getwd()}

\item{walltime}{PARAM_DESCRIPTION, Default: NULL}

\item{project}{PARAM_DESCRIPTION, Default: NULL}

\item{nodes}{PARAM_DESCRIPTION, Default: NULL}

\item{ppn}{PARAM_DESCRIPTION, Default: NULL}

\item{email}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
